feat: Dify AI 深度集成 + 表格填写文档查询 + Vault 增强 + 前端上传重构
【AI 引擎】
- 新增 Dify Workflow 通用客户端 (DifyWorkflowClient) 与配置体系 (DifyProperties)
- 实现 DifyAiOrganizeServiceImpl：自动规整服务对接 Dify Workflow
- 实现 DifyAiFormServiceImpl：表格填写助手对接 Dify Workflow（字段提取 + 文档生成）
- 新增 Dify 输入输出 DTO：DifyOrganizeOutput / DifyFormExtractOutput / DifyFormGenerateOutput
- application.yml 新增多 Workflow 配置块（organize / form-extract / form-generate）
- 保留 Mock 实现作为 fallback，支持 enabled 开关切换

【表格填写助手 - 文档查询】
- 新增 FormDocQueryController：/lingdoc/ai/form/query-docs
- 新增 IFormDocQueryService / FormDocQueryServiceImpl：按字段名从 Vault 检索相关参考文档
- 新增 FormDocQueryRequest / FormDocQueryResponse / FormFieldUpdateRequest
- 支持 Dify HTTP Request 节点调用，JWT 鉴权 + Vault 上下文透传

【Vault 模块增强】
- VaultController / VaultServiceImpl 新增接口与业务逻辑
- LingdocFileIndexMapper / LingdocTagBindingMapper 新增按 vault 查询方法
- VaultFileTree.vue 组件增强（交互与展示优化）
- VaultDbInitializer 初始化逻辑优化

【前端重构】
- 删除旧上传页面 src/views/lingdoc/upload/index.vue
- 新增 src/views/lingdoc/file-upload/index.vue：拖拽上传 + AI 自动规整引导
- 表单页面 (form/index.vue) 与 Vault 页面 (vault/index.vue) 同步调整
- API 层 upload.js / vault.js 接口更新

【文档】
- 新增 5 份 Dify 集成开发规范（12~15、17 号文档）
- 更新自动规整后端实现计划

【工程化】
- 删除临时文件 (.AGENTS.md.swp / temp_plan_writer.py / temp_write.py)
- ry.bat 大幅重构，package.bat / run.bat 同步更新
- logback.xml 统一控制台日志编码为 UTF-8
- SQL 脚本（菜单、初始化）同步更新
- 新增单元测试：FormDocQueryServiceImplTest / DifyAiFormServiceImplTest / DifyAiOrganizeServiceImplTest
- 新增 test/organize-samples/ 测试样本集

【其他】
- LingdocInboxServiceImpl 收件箱逻辑增强
- 各 Controller 参数与路径微调
- vite.config.js 配置微调

+++ ruoyi-server/ruoyi-system/src/main/java/com/ruoyi/system/domain/lingdoc/LingdocInbox.java
@@ -54,7 +54,7 @@ public class LingdocInbox extends BaseEntity
     private BigDecimal confidence;
 
     /** Token消耗 */
-    private Integer tokenCost;
+    private Long tokenCost;
 
     /** 错误信息 */
     private String errorMsg;
@@ -189,12 +189,12 @@ public class LingdocInbox extends BaseEntity
         this.confidence = confidence;
     }
 
-    public Integer getTokenCost()
+    public Long getTokenCost()
     {
         return tokenCost;
     }
 
-    public void setTokenCost(Integer tokenCost)
+    public void setTokenCost(Long tokenCost)
     {
         this.tokenCost = tokenCost;
     }
